+++ ios/history.md
@@ -3,6 +3,7 @@
 ## 11.0 alpha
 * Replaced deprecated calls to UIAlertView and cleaned up extraneous blank buttons (#1002)
 * Bookmark add button is enabled only when title/url fields have text (#1073)
+* Removed deprecated code to to support keyman:// scheme for ad-hoc distribution (#1160)
 
 ## 2018-07-06 10.0.203 stable
 * Fixes an issue where a keyboard with varying row counts in different layers could crash (#1055)